tests(e2e-tests): extract run-e2e-ci and cypress commands into test-utils
