+++ packages/e2e-tests/next-app/cypress/support/index.ts
@@ -13,4 +13,5 @@
 // https://on.cypress.io/configuration
 // ***********************************************************
 
+import "../../../test-utils/cypress/custom-commands";
 import "./commands";
