+++ packages/e2e-tests/test-utils/scripts/run-e2e-ci.ts
@@ -1,13 +1,15 @@
 #!/usr/bin/env node
 
+// Script to run e2e tests in a CI environment
+
 import fetch from "node-fetch";
 import { v4 as uuidv4 } from "uuid";
 import { execSync } from "child_process";
-import AWS from "aws-sdk";
-import fs from "fs";
+import * as AWS from "aws-sdk";
+import * as fs from "fs";
 
 // Next.js build ID follows a certain pattern
-const regex = /"buildId":"([a-zA-Z0-9_-]+)"/;
+const buildIdRegex = /"buildId":"([a-zA-Z0-9_-]+)"/;
 
 // AWS clients
 const cloudfront = new AWS.CloudFront();
@@ -17,7 +19,11 @@ const waitTimeout = parseInt(process.env["WAIT_TIMEOUT"] ?? "600");
 
 // Constants
 const deploymentBucketName = "serverless-next-js-e2e-test"; // For saving .serverless state
-const appName = "next-app";
+const appName = process.env["APP_NAME"]; // app name to store in deployment bucket. Choose a unique name per test app.
+
+if (!appName) {
+  throw new Error("Please set the APP_NAME environment variable.");
+}
 
 // To ensure cleanup doesn't happen more than once
 let alreadyCleaned = false;
@@ -49,7 +55,7 @@ async function checkWebAppBuildId(
 
       if (response.status >= 200 && response.status < 400) {
         const html = await response.text();
-        const matches = regex.exec(html);
+        const matches = buildIdRegex.exec(html);
 
         // Found match in actual buildId and expected buildId
         if (matches && matches.length > 0 && matches[1] === buildId) {
